Trait bmw_crypt::RngCompatExt

pub trait RngCompatExt: RngCore {
    type Wrapper: RngCore + OldRngCore;

    fn rng_compat(self) -> Self::Wrapper;
}

Extension trait for the current versions of RngCore; adds a compatibility-wrapper function.

Required Associated Types

type Wrapper: RngCore + OldRngCore

Wrapper type returned by this trait.

Required Methods

fn rng_compat(self) -> Self::Wrapper

Return a version of this Rng that can be used with older versions of the rand_core and rand libraries, as well as the current version.